Unit Tests: conditionally skip a test if required file isn't found.
This test was failing becuase travis is unable to find the file for some reason.  This same test passes in vvv.
This change simply checks for the existence, and skips if not found.

The error it was reporting was:

`filemtime(): stat failed for /tmp/wordpress/src/wp-content/plugins/jetpack/css/jetpack.css`