REST API: Sites: add headstart_is_fresh to site options data
This is a value that determines if Headstart considers the site to be "fresh" or not, meaning if it has modified content since the last run of Headstart. This can be used t o determine if Headstart could run again. It's not actually a site option (though we'll pretend it is here), but is derived from the blog stickers on the site.

Merges r137127-wpcom.