Advanced SEO: Add per-post custom meta description
Part of T473
Closes D2049

Adds the ability to save a new metadata field to post objects.

Previously, all descriptions and titles given to search engines were generated
automatically inside of WordPress. Now, this and a new title metadata field
will be available to certain users to customize how that information will be
generated.

Modifying this infomration currently requires a business plan subscription.

Props to @vindl for doing all the work on this patch.

Merges r137590-wpcom.

Conflicts:
	json-endpoints/class.wpcom-json-api-update-post-endpoint.php