Jetpack Sync: Call spawn_cron after scheduling full sync
In order to not block the API, we schedule a cron event to start a full sync. While this makes sense so that we are not slowing down the API request, it has the side effect of the UI not being as responsive.

To get around this, we can call spawn_cron() immediately after scheduling our full sync which will fire off an HTTP request and get our full sync started.

Note: spawn_cron is throttled to only run once per minute.