Jetpack Sync: Add is_scheduled to GET /sites/$site/sync/status endpoint
When we trigger a full sync through the API, we do this via cron so that we do not block the API while enqueuing items.

While scheduling a cron event handles the issue of blocking the API gracefully, it can take several minutes for a scheduled event to run. This means that a user could trigger a sync via the API, and not begin seeing any progress from that sync until several minutes later.

To address this, we can add an is_schedule property to the response so that the UI can respond accordingly. This can be as simple as showing "Sync is scheduled text, check back soon" text, or can be handled in another way.