Sync: Preserve the home_url, site_url and network_url
This PR tries to make the urls more stable in the scheme sense. Since
the scheme can change depending on what url the user uses to send the
data from on shutdown.

This PR tries to fix this by matching the schema of the url more
closely to the url option defined in the general settings of the site.