Add VaultPress and Jetpack 4.1 compatibility
Since we though that we would ship Jetpack 4.1 with the new sync we
added code to VaultPress to turn off the old syncing in version 1.8.3.
This code adds VaultPress compatibility. Which is VaultPress is
installed sends same data as before.

This code doesn't exist in new Version of Jetpack. But is there for
version 4.1 to make sure that sync continuous to work as expected.
Without replying on a specific version of VaultPress.