Social Links: Enable removal of disconnected services
If link is present in options but the connection is no longer available or invalid (e.g. app-scoped token), the user will not be presented with the option to unselect a connected service.

These changes iterate over all available services and checks for presence of link option. While we can't show the display name, we can still allow them to unselect the option.

See: https://[private link]

Merges r138651-wpcom.