Advanced SEO: Represent nonexistent custom title with empty string
Closes D2247

Previously, when updating the custom SEO title formats via an API call
(`/site/settings`) and only providing a partial list of updated format types,
the other types were being unset.

Now, only those types of formats (front page, post, pages, etc...) will be
changed when updating via an API call. The others will be left as they were.
The way to properly unset an existing title format is to pass an empty string
set for that type.

Props to @vindl for doing the work for this patch.

Merges r138837-wpcom.