Publicize: Jetpack_PostImages: Fake the full image if no image is return by wp_get_attachment_image_src.
This fixes the issue for new jetpack sync (4.2) where the publicised tweets were not including the featured image of a post in the tweet. Since in the new sync implementation we don't save the post meta information about the featured image  in the post meta. Now we sync data about the featured image and the attachment info as well.

Merges r139875-wpcom.