Fix issue where users wasn't able to login.
This fixed an issue reported by users where they were not able to login
if they had W3TC installed and all the caching enabled.

The issue was that we seem to be overwriting the user object with a
user object that doesn't contain the password has.

This PR fixes it in 2 ways. It delays the sanitisation of users when
meta data capabilities sync might happen. And it clones a new user
object before modifying it so that the user object stored in cache
doesn't get modified.

- We added a test that shows that we don't sync random user meta.
- We were not able to write a good tests showing the modification of
the user object in cache.