Advanced SEO: Display custom title formats
Closes D2341
Closes T473

Displays custom title instead of the default one, provided that current
site has a business plan, and that the user has defined a custom title
format for current page type.

In this revision, caching was added to return value of has_business_plan
check, in order to avoid performance issues caused by it when determining
whether the default title should be replaced.

Props to @vindl for doing the work on this revision

Merges r140462-wpcom.