Fix PHP warnings when user subscribes to comments
Reported in #3339, PHP warnings are generated when a user subscribes to comments, and the cookie is set with a null value instead of the expected post ID, making the cookie useless since the check in `comment_subscribe_init()` will never pass.

This patch fixes the reference to the post ID, and it saves the "Subscribe to all comments" value for each post separately. Previously, if a user subscribed to comments on post 1, then commented on post 2 and did not subscribe to comments on it, the checkbox on post 1 would remain unchecked. Now, both checkboxes will have the appropriate state.

This will increase the number of cookies set for users that subscribe to comments on many individual posts. We may wish to revisit the ~1-year cookie expiration limit.