Publicize API Endpoint: A new API endpoint that allow you to re publicize your posts.  (2/2)
It uses the new publicize classes.
Which are updated so that we can pass in values to bypass the data that is currently stored in post_meta.

This work was part of the 2016 GM Project described here.
https://[private link]

Code review can be found here https://[private link]
Previous commit is https://[private link]

Merges r142558-wpcom.