Add `get-media` and `update-media` endpoints `v1.2`
Summary:
The goal of this new version of the update media endpoint is achieving update of the attached file according to of media ID.
In terms of the API usage the endpoint now accepts `media` and `media_url` so if one of them is sent the attached file will be updated.

Once the image is updated the object response gives us the `revision_history` property which is an array of all previous images. It will very useful to rollback to a previous image, show history track, etc.

Merges r143528-wpcom.