Do not sync non public post types filtered post content and excerpt. (#5288)
* Do not sync non public post types filtered post content and excerpt.

The reason why we don't want to sync that content is that it takes
resources to generate and the filtered content but likely good that we
would use it is very small. Since the post type is internal anyway.

* Sync: Make sure that we don't sync non existant post types

Post types that are not in the registed any more could still be in the
DB and hence produce sync errors this way when we send data.

This commit tries to fix that.