posts/new: Updated v1.1 & 1.2 endpoints to use media IDs in new posts gallery tags
This slightly modifies existing behavior when media is attached to a new post. Previously a gallery shortcode was added to post content without any media IDs associated with it. This works for most things but breaks things like Publicize and the Calypso editor. This commit changes existing endpoints to associate media IDs with the shortcode.

The primary reason for fixing this is when sharing a photo from the iOS app Publicize was picking the wrong image.

https://[private link]

Merges r144544-wpcom.