Auto bump w.js version weekly
We sometimes make changes to our analytics JS file `w.js` which is outside of the Jetpack upgrade cycle. In order for clients to load the latest file we should just force a new version weekly. This way we still take advantage of caching as the JS file is only loaded once a week in the worst case scenario. This is similar to how we handle the stats script in the stats module:

https://github.com/Automattic/jetpack/blob/6d6681e0df12402010d3331cb6e9dc5c674b5dfe/modules/stats.php#L125