Fix Carousel not starting for single image after its link was edited
Before the fix Carousel for single images was dependent on attribute rel
of 'a' tag wrapping the 'img' tag. This attribute is set differently when
image is added to post and when its link is edited in Media Settings modal.

After the fix this attribute is not used - additional data-permalink
attribute was introduced to img tag that allows easy diffrentiation
whether image links to 'Media File' or 'Attachment Page'

Fixes: #5431