Add filter to bypass check_login_ability checks. (#5369)
* Add filter to bypass check_login_ability checks.

Adds a filter to bypass the `check_login_ability()` checks.

The use case here is our own function that verifies users. When it returns true, we know the login should not be blocked and we want to bypass the JPP checks.

An alternative way to implement this would be to filter the IP list to return an array with the current remote addr if we want to avoid being blocked. The problem is this is not super straightforward as you have to filter the underlying Jetpack option.

* Add docblock for jpp_allow_login