[contact-form $attributes] now ok in do_shortcode (#3188)
* [contact-form $attributes] now ok in do_shortcode

Previously, when using the [contact-form] shortcode in do_shortcode situations like a page template, the attributes like 'to' and 'subject' were being dropped. $attributes and all its variations aren't available by the time we get to processing the submission, so I had to rewrite it as a string earlier in the process, in store_shortcode(), when we still have access to $attributes. If someone knows how to get access to the shortcode attributes in process_form_submission() then we could save a few lines of code and keep attributes as an array.

* Keep $attributes as array until it's needed, perform checkings on feedback meta data returned, filter empty values.