Manage: Make a "default" module. (#5690)
Manage is already activated during the connection flow, and we've already removed the UI for deactivating it since 4.1

Since this was not listed as a default module, it was causing some bugginess when testing Jump Start after "Reset Options".  Since the modules were reset to default, manage was deactivated and then on connection it was actiavted again, thus confusing Jump Start into thinking a user had taken action so it would not show.