WPCom child themes: Template: will be suffixed with -wpcom
A WPCom child theme's style.css file contains a `Template:` prop
pointing to a WPCom parent theme. However, this prop doesn't contain the
`-wpcom` suffix. As all the WPCom themes are suffixed with `-wpcom`, we
need to manually add this suffix to all the WPCom child themes.

In this commit, I honor the above assumption instead of adding the
`-wpcom` suffix in the code.