Export endpoint - fixed args for partial exports
Partial exports have a set of parameters that define what parts of the site should be exported.
setup_args is a function that translates one set of parameters received from the API request to another set delivered to the .org exporter.
This function was fixed and is now working, supporting all partial exports.