Apps Landing: Adding CTA at top to download the apps
Apps Landing: Adding CTA at top to download the apps. Previously the
links to download the apps were only way down on the page. This CTA
takes the users to apps.wordpress.com instead of lower on the same
page. This is due to a current restriction in our react setup where
creating anchors to link to locations on the same page is not possible.