REST API: SAL: Move icon logic to Jetpack base, leverage core site icon option consistently
The Jetpack Site Icon module was deprecated in 3.6.1. The transition logic deactivates the module for most sites, so the jetpack_site_icon_url function usually does not pass the existence check for most Jetpack sites.

In any case, we plan toward moving to treating Site Icon as the canonical source of site's icon. The changes herein check for this option, falling back to the Jetpack-defined icon option if unset.

This improves consistency between Jetpack shadow sites and the remote site. In my case, /me/sites (shadow site) will show the icon property for my Jetpack site, whereas /sites/%s does not (because the module is not active on the site).

Eventually the goal is to move get_icon to the SAL base site class, dropping Blavatar as a dependency for WordPress.com sites.

Differential Revision: https://[private link]

Merges r148273-wpcom.