Site Icon: Remove Blavatar and return as though cleared when unsetting site icon
Blavatar deprecation: If trying to clear a site icon when an option is not
already configured, we still want to remove the Blavatar if one exists. The
mechanism already exists to remove Blavatar if (un)assigning a site icon,
but the action will not be triggered if the value does not change. Checking
here allows us to guarantee the site's icon, whether Blavatar or Site Icon,
will be removed upon passing a falsey value.

See: https://github.com/Automattic/wp-calypso/pull/10387 (Caveats)

Merges r148386-wpcom.