Photon: Use `wp_get_upload_dir`
In WP 4.5, wp_get_upload_dir was added as a helper function that provides the same functionality of `wp_upload_dir` without opting to create the directory if it doesn't exist. Since we are not writing files, we don't need to create any directories nor check if they exist, so we can be a tad more performant.