Twitter Widget: add noscrollbar option back.
The option was removed in #4198, but it is still supported by Twitter:
https://dev.twitter.com/web/embedded-timelines#customization

To be be able to hide the scrollbar, you should not specify a number of tweets in the widget.
If you do, the `data-tweet-limit` attribute will be added to the widget and the timeline will automatically adjust its height to fit the
specified number of Tweets. That's most likely why we thought the option didn't work when testing in #4198.

Reported here:
https://twitter.com/LorrieW_Art/status/822005258096406529