Make close notice icon clickable (#6182)
* Make close notice icon clickable

Bump the z-index of the x close link to make sure that the connect
banner is closable. Since currently you can't click on the link to
close the banner.

* Make the banner more readable

By making sure that the icons are always aligned at the same width and
that text lines are not super long.