Make it harder to use file locking because it's not really needed.
The plugin writes to files with random filenames and then renames those
files to their final cached filename which is an atomic action that
can't be interupted. The only reason to have this coarse locking is if
your server is so underpowered it can't cope with multiple processes
writing to files at the same time. This will slow down how fast new
files are cached significantly.
This patch changes the language on the settings page warning the user
that if will slow down their website.
The cosnstant "WPSC_DISABLE_LOCKING" was used to hide the setting, but now
it's checked in each function that implements file locking.
sem_remove() seems to cause problems using semaphores so it is disabled
unless the constant "WPSC_REMOVE_SEMAPHORE" is defined.