Store legacy cache files in the supercache directory.
If we store the legacy cache files in the supercache directory it will
be much easier to maintain them as they'll be in a directory structure
that reflects the URLs served by the site, just like supercache files.
This patch checks the original blogs_cache directory first but any new
cache files will be created in the supercache directory.