Rename the Top Posts stylesheet to widget-grid-and-list for sharing
The style is used by other widgets with a grid of items (Blogs I Follow). This
change moves the stylesheet into a common location where it can be registered
and enqueued by any widget that requires it. The name is also updated in the
Jetpack CSS concatenation and gulpfile concatenation lists.