Fix: Use the site_icon id instead of the url (#6303)
When $image_url is set to a photon image we are not able to deremine
the $image_id. Instead we should use the option that stores the ID
instead.

This fixes the issue when we show the default image instead of the
site icon as the open graph main image. When photon is enabled and the
site icon is set.