Widget Visibility: Rewrite the way that the major/minor rules lists are generated. (#6333)
* Widget Visibility: Rewrite the way that the major/minor rules lists are generated to save bandwidth and memory.

In situations where a site has hundreds of widgets each with multiple visibility rules, the widget admin pagesize can grow out of control, causing out-of-memory errors and issues with loading the customizer.

This commit rewrites the way that the admin UI is generated, moving from hardcoding everything server-side to providing a single set of all the possible visibility options that are then dynamically generated client-side when they're needed.

This also has the happy side-effect of eliminating the need for the AJAX calls when changing the major rule type. It also fixes some inconsistencies with how the "Include children" checkbox was processed.

* Address JSHint issues.

Address some comparison and variable declaration issues, mostly by eliminating them alltogether by using a simpler method of setting the value of a select input.