Fix Content-Type header for OPTIONS preflight request
Summary:
[108139] introduced an explicit 200 response for preflight requests.

Instead of sending an empty body as intended, though, it sends `'""'` since the
Content-Type header is set to `'plain/text'` instead of `'text/plain'`. This
causes the API to send the empty string JSON-encoded (`'""'`) as the response
body.

Change the Content-Type to `'text/plain'`.

Test Plan: Sure.

Reviewers: jnylen0

Reviewed By: jnylen0

Subscribers: jnylen0

Differential Revision: https://[private link]

Merges r150936-wpcom.