Jetpack Sync: Consolidate nearly duplicate sync and full_sync methods
The Jetpack sync and full_sync methods were nearly exact duplicates of each other, so I kept the "interface" to the two methods for the sake of any clients calling them, but delegated to a single helper method that behaves one way or the other depending on the identifying parameter passed to it by one of the original methods. Less duplicate code means less to grok, less to maintain, and less opportunity for oversights keeping the methods in sync if there are any changes to them.