Added building to the lint script.
Sometimes it happens that the Travis build passes when the build actually fails. It happens because we weren't building the actual bundle before, but this commit introduces a build command before actual linting. That will make sure that Travis builds fail if the project doesn't build successfully or that jshint fails.