Fixes published-post autosave-previews in calypso (#6530)
* Fixes published-post autosave-previews in calypso

The wpcom endpoint cannot predict whether the user is logged in or not on the remote site. So it cannot pass the appropriate nonce that gets verified in `/wp-includes/revision.php`. This patch detects the frame-nonce parameter for a caypso preview falls back to the existing frame-nonce validation when it exists.

* fixes spacing and adds textdomain for translation