Fix calendar reader timezone offsets
This revision aims at fixing an issue in the `iCalendarReader` class
The function that applies timezone offsets calculates the time difference for timezones using the current date instead of the date thats actually associated with the event. This leads to Daylight Savings Time confusions if today is standard time and the event occurs in daylight savings time. Today (3-8-2017) an event for 3-15-2017 6:30pm will show up as 3-15-2017 5:30pm

For effected users see:
  - https://[private link]
  - https://[private link]
  - https://[private link]

Merges r152463-wpcom.