SEO Tools: Replace blog sticker used for business plan check
As part of Premium in Premium project, unlimited premium themes will
now be offered to all sites that have a Premium plan too. It was decided
to implement this by adding unlimited-premium-themes sticker to those
sites. That means that we can no longer use this sticker to reliably determine
the presence of the business plan. Because of that we are switching to a new
one that was introduced as a replacement (business-plan).

Differential Revision: https://[private link]

Merges r152594-wpcom.