Open Graph: do not return images smaller than required size.
Fixes #2284

When using `jetpack_og_get_image`, we specify a minimum width and height (200px).
Those 2 values are then used by `Jetpack_PostImages::get_images()`. `get_images` looks for images following different methods, and we
applied the minimum width and height to most of them, thus excluding images that were too small. However, we did not apply the values for
the last fallback used to look for images in a post: `from_html`.

This commit implements the 2 width and height requirements to `from_html`. To do so, I switched from a regex match to using `DOMDocument`
to fetch the different image tags and their width and height values.