Added a delete cache endpoint to the rest api
Delete the cache of the current blog by sending an emtpy POST to /delete
Send a JSON encoded post ID to delete the cache for that post/page.
Added a new function wpsc_delete_post_cache() to delete a single post,
given a post_id.
And fixed a bug in wp_cache_clean_cache() where $supercachedir is not
set. (shouldn't happen too often)