Allow HTTP PUT data to be validated by Jetpack_Signature
HTTP requests made by functions like `Jetpack_Client::remote_request()`
generate signed body hashes for the request content. If the data is sent
via the `PUT` method, however, it was ignored, causing a hash comparison
failure on the server. This change allows `PUT` data to be validated as
well.

Just like the `POST` data, the `PUT` data is taken from the raw
url-encoded input (I don't understand why), but if IS_WPCOM, we assume
the data is JSON-encoded and use the decoded version.