Jetpack Related Posts: Fix fatal when the exclusion parameter is not a string.
Some requests for related posts are using a malformed URL where the `relatedposts_exclude` parameter is of type array, not of type string, causing `explode()` to fatal.

This change checks the type of the `relatedpost_exclude` parameter, and will perform the original explode if it is a string; otherwise, we sanitize the array for use as `$excludes`.

See https://a8c.slack.com/archives/C4N88L95W/p1490977986140965

Merges r153923-wpcom.