Related Posts: Fix fatal when the exclusion parameter is not a string
Some requests for related posts are using a malformed URL where the
`relatedposts_exclude` parameter is of type array, not of type string, causing
`explode()` to fatal.

This change checks the type of the `relatedpost_exclude` parameter, and will
perform the original `explode()` if it is a string; or extract the values if
it's an array. We sanitize the values to ensure they are valid ints for use as
`$excludes`.

Props @jmdodd