Contact Form: Set message content type to restore compat with plugins
Some plugins override `wp_mail()` and replace PHPMailer with other libraries, like Zend Mail. Previously, we were relying on PHPMailer to correctly set the content type, but that never happens when other libraries are used. Because of that, the message can be sent out with a `text/plain` type, but contain HTML, making it difficult for users to read.

All plugins should respect the value of `wp_mail_content_type`, though, so this is a much more reliable method of setting the type to `text/html`.

Props georgestephanis, jeherve
Fixes #7003