SEO Tools: Allow empty string to be passed as title format (#7044)
* SEO Tools: Allow empty string to be passed as title format

Previously, only arrays were considered as valid title formats, and
in case of deletion we needed to provide empty array. Now, we are also
allowing empty string to be passed in that case instead. This is needed
to resolve the bug that prevented title deletion in Calypso.

GitHub issue: https://github.com/Automattic/wp-calypso/issues/10003

* yoda conditional