When rebuilding, delete unmodified base files at the end of the process
When a page update is triggered and "needs rebuild" feature is turned on
then an issue occurs if more than one type of cached visitor has
visited. For example - desktop and mobile browsers, with mobile support
on.
The next visitor type will refresh the rebuild files for their type, but
not the other(s).

This patch records the file modifcation times of each cache file and
deletes any that are unmodified at the end of the process. Slightly out
of date cache files for "other" types will be deleted.

A small change to the "do_rebuild_list" variable means that cache
files are deleted when not caching the current page ONLY if a rebuild
was triggered.