Add a defined() check on WPLOCKDOWN and load config file for settings
Load the plugin config file in the settings function so that the
activity of the plugin doesn't change any setting while getting the
settings through rest api.
The config file may have a constant WPLOCKDOWN defined so put a
defined() check in front of it if that check isn't there already, to
stop PHP notices appearing.