Shortcodes: fix recipe time display to be compatible with Schema
We previously displayed the total recipe time as the user provided it in the shortcode.
While that works on the site, that format isn't compatible with Schema.org.
Search engines relying on Schema.org to display recipe previews expect the recipe time to use ISO 8601 formatted duration
interval.

This commit does 2 things:
- It adds a new shared function allowing you to convert any user-inputted duration into ISO 8601 formatted duration.
- It uses the shared function to output a new `time` tag in the recipe output, where that formatted duration is added as a
datetime attribute.