Improve the performance of the Follow Button widget.
The follow button can be seriously slowly rendered in dotorg sites
due to one of the batched requests, which heads for sites/http://foo.com.

This PR adds `data-blog-name` to the button element,
so the iframed widget can omit the slowest one of the bached requests.
In my test, the widget is rendered 3x faster
when it doesn't need to show the subscribers counter.

Note: It's still slow when the counter is necessary.

You need D5561-code to see the improvement.