Backups/scan card: Fix broken status for personal plan.
This fixes an issue when we were showing the wrong text on personal plan.  It also re-writes the logic to show the content to make it more readable.

It also fixes an issue where we were not showing the `set up` button on professional plans if scan was not enabled for some reason.