API: Add new endpoint version for listing posts
This introduces a new parameter to the "list posts" endpoint (it creates a new version of that endpoint). That param is `exclude_private_types` and can only be used in conjunction with `type=any`, as noted in new docs. This is useful for cases where we want to show the user only posts that their visitors can access (eg. include posts, pages, portfolio, but don't include feedback and other private types).

See Phabricator Diff 4766 for more info
API change required for https://github.com/Automattic/wp-calypso/pull/11887

This commit syncs r156614-wpcom.