Action: `jetpack_user_authorized`
We currently do this action before we save the user's token to the database.

Since this action effects JetpackTracking::track_user_linked which requires
a user token, we should only do `jetpack_user_authorized` after the token is saved.

This PR ensures we do `jetpack_user_authorized` after a user token is saved to
the database.