Only make GET non REST API requests anonymous
The plugin allows the user to make every request anonymous to make the
site more cacheable but unfortunately that caused problems with services
that interacted with the site in an automated manner like REST API
clients. This fixes that problem by removing the login cookies only if
the request is a GET request or isn't a REST API request.