Process all unloaded widgets instead of just the first one
This was a remnant of previous setTimout approach in queue handler
that picked up only the first unloaded widget. This worked fine before,
but now that the execution of queue handler is not pooled so frequently,
it caused issues with some widgets not being loaded after scrolling.

With this patch we will correctly load all of the unloaded widgets that
are visible.

Also removed some old obsolete code.