Update the .htaccess rules safely without bringing down the site
mod_rewrite rules have to be updated to use Export mode but if done
incorrectly the site can be taken offline. This code makes a backup of
the .htaccess file, grabs a copy of the front page, updates the rules,
grabs the front page again and if they are not the same then it restores
the backup.
TODO: more error reporting as there are so many checks.