Sync: WooCommerce - Ensure we sync the order_id for order_items in older WooCommerce versions.
Prior to WooCommerce 3.0 the second argument to the woocommerce_new_order_item action only contained the order_item_name and order_item_type information and not the order_id.

This meant that incremental syncing of this data would loose the relevant order_ids and the products would not be correctly associated with the orders.

Relevant code in previous major version: https://github.com/woocommerce/woocommerce/blob/2.6.14/includes/wc-order-functions.php#L515