Contact Form: Escape `@` in email addresses to avoid transport errors
`@` symbols in email addresses should be escaped, according to RFC 5322. Most MTAs won't complain if it's not, but some will refuse to send the message. Most reports have been from sites hosted on SiteGround, but a few others may be affected.

This is a work in progress. I think it fixes 90% of the issue, but it needs some cleaning up and testing.

See #7276