fix HTTP status from math-fallback.php
I've been seeing the occasional HTTP 500 status code in my web server
log for a long time in response to requests such as 'POST
/wp-login.php'. I finally took the time to track it down to a bug in the
math-fallback.php module. The problem is that the invocation of
`wp_die()` is incorrect. This change fixes that bug. I now see the
expected 401 status in my logs.