Add Comment Likes module (#7175)
* WIP: Attempt to bring back the old comment likes iframe to working condition

* Move comment likes code to new comment like module

props @roundhill

* Clean up unused comment likes code

* Refactor function name to comply with coding standards

* Remove obsolete code

* Don't load like widget iframes until they are scrolled into view

* Perform only one batch request for all like widgets on page

* Unload comment like iFrame widgets that are scrolled out of view.

* Add styles for comment likes loading placeholder

* Refactor iFrame unloading

* Fix like and unlike actions

* Preload widgets that are close to being viewed

* Bump likes widget cacheBuster version

* Update module DNS prefetch URLs

* Load all widgets immediately on pages containing up to ten comments

* Refactor code to make JSHint happy

* Match widget appearance with WP.com version

* Make sure that master iframe is always loaded when needed

* Settings UI: don't falsely toggle comment likes when toggling post likes

* Attempted a few fixes to get the loading indicator and the like button iframe to match up heights.

Also added the star noticon start to help with the transition, seems easier on the eyes.

* Matching up styles to latest version of iframe.

* Code refactor

* Align widget placeholder text with actual widget

* Improve unloading of widgets that are scrolled out of view

* Factor out setTimeout calls from queue handler function

These were causing queue handler to execute too much, causing
high CPU load when there is a large number of widgets on page.

* Limit the size of initial batch request

Previously, all unloaded widgets were included in initial batch request,
which caused the API call to timeout when too many comments are
present. This limits the size of the initial batch to the widgets that are
scrolled into view.

* Remove jQuery inView plugin that was causing performance issues

It was attaching event to all of the unloaded widgets and was generating
a lot of CPU load when there are many of them on the page.

* Limit processing to unloaded widgets that are scrolled into view

* Factor out widget iframe loading into separate function

* Process all unloaded widgets instead of just the first one

This was a remnant of previous setTimout approach in queue handler
that picked up only the first unloaded widget. This worked fine before,
but now that the execution of queue handler is not pooled so frequently,
it caused issues with some widgets not being loaded after scrolling.

With this patch we will correctly load all of the unloaded widgets that
are visible.

Also removed some old obsolete code.

* Move module cart to a separate component

Also updated module info.

* Code refactor

* Fix likes module activation bug

In cases when comment likes module was enabled, try to enable the
post likes module failed. This was caused by fatal error due to the
attempted redeclaration of jetpack likes master iframe.

* Revert change that was breaking like widget unloading

* Clean up remaining usages of jQuery inview plugin

* Always use https protocol for widgets.wp.com

* Add phpDoc for comment likes button content filter

* Add comment like counts to wp-admin comments page

Adds a separate column on comments page that will show comment
likes in a similar way that post likes work.

* Comment Likes should obey individual post "Show Likes" settings

Make Comment likes respect the global and individual rules for displaying
as Post likes do. Part of the Likes module code was factored out
into a separate class in order to reuse it more easily in Comment likes module.

* Fix failing unit tests

* Refactor code

* Move comment likes activation to Discussion tab

* Comment likes: consolidate calls over dom element, fix whitespaces, use $ instead of jQuery.