Add new Social Icons widget and deprecate old Social Media Icons widget
To keep in line with WPCOM, I’d like to add the new Social Icons
widget. You can add this widget to any site and you should have a new
widget as seen on its documentation page
(https://en.support.wordpress.com/widgets/social-media-icons-widget/).

The old widget should become unavailable to sites not using it. If they
do then we add a “(Deprecated)” to the label.

In `social-icons.php`, L72, we use the same `social-menu.svg` file as
the one used for the Social Menu.