API: Add password and status to list of keys to always render
Follow on from r156885, this adds the password and status fields to the list of fields that should always render for post objects. These are needed because downstream code looks at the password and status fields to determine if a post should be shown or not. When a request is made and these fields were not included, those checks would always fail and posts would not be returned.

This commit syncs r159288-wpcom.