Fix edit view render for Firefox.
We were writing the contents too quickly after adding the empty iframe
to the dom -- so when load ran, the contents got wiped out.

Instead let's do our stuff on the load event.

The downside is that there's a brief flash of unstyled content, will
look at addressing seperately -- as that's gloss, not functionality.

Props @petitphp for the find!

https://twitter.com/petitphp/status/889968381994160128