Don't cache unknown page types, fixes problems with sitemaps (#346)
* Don't cache unknown page types, fixes problems with sitemaps
Fixes 325
Sitemaps are usually generated by a plugin that exits immediately which
causes problems for caching as it doesn't know what type of page it is.
This code refuses to cache pages if the plugin doesn't know exactly what
type of page is being served. A debug log entry is also generated.

* Typo fixed in query check.

* Improve detection of sitemaps by using query_vars, props @stodorovic

Add X-Robots-Tag header used by sitemaps.
Fix $headers/$response mixup introduced in 1.5.0

* Use the is_feed from $wp_super_cache_query instead of is_feed().

The variable $wp_super_cache_query[ 'is_feed' ] is set when a sitemap is
detected as well as a feed making it more useful.