REST API: Set date when manually publishing scheduled post
When publishing scheduled posts directly from the `posts/scheduled/{site}` screen in Calypso, the API request to the update endpoint would only the `post_status` field. When the endpoint passed that data on to `wp_insert_post()`, it would be considered invalid input, because it the existing scheduled `post_date_gmt` field was used, and that date is in the future.

Core doesn't have a flow similar to Calypso's, so it never encounters this condition normally, and it expects posts to have dates in the past when they're being published. In response to the invalid input, `wp_insert_post()` would override the `post_status` and set it back to `future`, so the post would never be published.

The solution to that is to explicitly set the date to the current timestamp before calling `wp_insert_post()`.

Fixes https://github.com/Automattic/wp-calypso/issues/4330
Differential Revision: https://[private link]

Merges r160864-wpcom.