SSO: Allow redirecting to JETPACK__API_BASE in production
When I initially added this logic, I only wanted to allow the change for sites that were in some form of non-production mode. But, this also has the effect of breaking the auto-authorize redirect for a production site if the the site has defined `JETPACK__API_BASE` to something other than the default.

This change should only affect SSO redirects, and as such, adding the extra URL parsing to requests should not be a performance burden.

To test:

- Checkout branch
- Connect site
- With a user other than the the master user, disconnect, then login via SSO
- After clicking "Log In" on WPCOM, you should end up the auto-authorize flow of JPC
- Try the previous two steps after setting `JETPACK__API_BASE` to your WPCOM sandbox

cc @gravityrail for review.