Rest-API: Introduce a bulk post deletion endpoint
Note: This commit only adds the new endpoint file. The remainder of D6769 will follow in a subsequent commit.

Summary:
This diff introduces a new bulk post deletion endpoint. The endpoint will be used in the future to enable bulk deletion of posts via Calypso. The endpoint follows the behavior of the current single-delete endpoint: if trash is enabled, the first call will send a post to the trash, and subsequent calls will permanently delete a post. If trash isn't enabled, any call will permanently delete a post.

Author: @jonathansadowski
Merges r160521-wpcom.